# Add slim endpoints for entities to speed up filters