--- a/pkg/api/resolver_query_find_studio.go
+++ b/pkg/api/resolver_query_find_studio.go
@@ -25,3 +25,8 @@ func (r *queryResolver) AllStudios(ctx context.Context) ([]*models.Studio, error
 	qb := models.NewStudioQueryBuilder()
 	return qb.All()
 }
+
+func (r *queryResolver) AllStudiosSlim(ctx context.Context) ([]*models.Studio, error) {
+	qb := models.NewStudioQueryBuilder()
+	return qb.AllSlim()
+}