--- a/pkg/api/resolver_query_find_tag.go
+++ b/pkg/api/resolver_query_find_tag.go
@@ -16,3 +16,8 @@ func (r *queryResolver) AllTags(ctx context.Context) ([]*models.Tag, error) {
 	qb := models.NewTagQueryBuilder()
 	return qb.All()
 }
+
+func (r *queryResolver) AllTagsSlim(ctx context.Context) ([]*models.Tag, error) {
+	qb := models.NewTagQueryBuilder()
+	return qb.AllSlim()
+}