--- a/pkg/database/migrations/7_performer_optimization.up.sql
+++ b/pkg/database/migrations/7_performer_optimization.up.sql
@@ -100,5 +100,3 @@ FROM `temp_old_performers_scenes`;
 
 DROP TABLE `temp_old_performers`;
 DROP TABLE `temp_old_performers_scenes`;
-
-VACUUM;