--- a/pkg/models/querybuilder_performer.go
+++ b/pkg/models/querybuilder_performer.go
@@ -106,7 +106,7 @@ func (qb *PerformerQueryBuilder) All() ([]*Performer, error) {
 }
 
 func (qb *PerformerQueryBuilder) AllSlim() ([]*Performer, error) {
-	return qb.queryPerformers("SELECT performers.id, performers.name, performers.gender FROM performers " + qb.getPerformerSort(nil), nil, nil)
+	return qb.queryPerformers("SELECT performers.id, performers.name, performers.gender FROM performers "+qb.getPerformerSort(nil), nil, nil)
 }
 
 func (qb *PerformerQueryBuilder) Query(performerFilter *PerformerFilterType, findFilter *FindFilterType) ([]*Performer, int) {