--- a/pkg/models/querybuilder_studio.go
+++ b/pkg/models/querybuilder_studio.go
@@ -93,6 +93,10 @@ func (qb *StudioQueryBuilder) All() ([]*Studio, error) {
 	return qb.queryStudios(selectAll("studios")+qb.getStudioSort(nil), nil, nil)
 }
 
+func (qb *StudioQueryBuilder) AllSlim() ([]*Studio, error) {
+	return qb.queryStudios("SELECT studios.id, studios.name FROM studios " + qb.getStudioSort(nil), nil, nil)
+}
+
 func (qb *StudioQueryBuilder) Query(findFilter *FindFilterType) ([]*Studio, int) {
 	if findFilter == nil {
 		findFilter = &FindFilterType{}