Should resolve issue #452 

In my library with 123 performers and 1MB-ish performer images it speeds up AllPerformersForFilters from 0.4s to 0.04s.  For 549 studios it speeds up from 0.08s to 0.02s.

I considered changing the dropdowns to fetch on demand, but considering that scenes can have arbitrary amount of performers attached, it would likely make page load significantly slower in many instances.