@WithoutPants 2020-04-13
Testing against a database with 1,968 performers, I'm not seeing any difference before and after this change. I'm seeing between 7.5 and 12 seconds for `allPerformersSlim` to come back.