@InfiniteTF 2020-04-13
@WithoutPants Try now. I've added a migration that moves the `image` column to the end of the table. Apparently selecting columns past blobs incurs a large penalty in sqlite. On my instance it reduces the allPerformers query further down to 0.01s, though it seems my dataset isn't large enough to really stress the database at this point. I'm guessing at a certain size it spills to disk, causing the crazy query times.