@bnkai 2020-04-13
I was getting consisntent speedup in the edit scene menu as stated above also. @WithoutPants is your db maybe in a spinning disk? ( my tests were in sata ssd and nvme ssd)