@bnkai 2020-04-13
With latest commit : 
Database located on a nvme ssd
Edit Scene Tab
Time for AllPerformers
500+ performers  
before 800ms after 15-20 ms
1k+ performers
before 1150ms  after 25-30 ms