@WithoutPants 2020-04-16
I'm getting consistent times of around 50ms for `allPerformersSlim` now. Looks good.